mediscount-github/requirements.txt
streamlit
pandas
geopy
requests
folium
streamlit-folium
python-dotenv
